chore: Extract into `CameraSession+Configuration`
